requirements.txt
python-jobspy
pandas
numpy==1.26.4
